requirements.txt
streamlit==1.39.0
pandas==2.2.3
plotly==5.24.1
python-dotenv==1.0.1
langchain==0.3.3
langchain-openai==0.2.2
langchain-experimental==0.3.2
tabulate==0.9.0
watchdog==5.0.3